requirements.txt
pytest~=7.1.2
PyYAML~=6.0